[Refatoração] - Ajustes de estrutura pastas - Refatoração de client.tsx     - Criação componentes reutilizáveis para os Cards de Carros e de Solicitações.     - Separação a lógica de solicitação e manipulação de dados em funções separadas.     - Utilização de funções assíncronas para melhorar a legibilidade e tratamento de erros.
