+++ code/frontend/package.json
@@ -16,6 +16,7 @@
     "axios": "^1.5.0",
     "eslint": "8.49.0",
     "eslint-config-next": "13.5.2",
+    "moment": "^2.29.4",
     "next": "13.5.2",
     "postcss": "8.4.30",
     "react": "18.2.0",
